Restrict contructors for Approx EachApprox UpToPhase
Previously these could be constructed with an argument of any type
and an instance would be created with this argument in the field `kw`.
For example `Approx(1e-10)`. But the argument is only useful if it is
keyword arguments of type `Pairs`. Trying to use  `Approx(1e-10)` will
immediately raise errors.

This commit disallows construction of `Approx` (and the others) with
anything other than type `Pairs`.

The version number is bumped to 1.0.0. There is no good reason to
keep releasing unstable versions.